backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
spacy>=3.7.0
requests>=2.31.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
mistralai>=0.4.0
